Moved old tests to the right location
