+++ smslibrary/src/androidTest/java/com/eis0/smslibrary/SMSCore_Tests.java
@@ -65,5 +65,4 @@ public class SMSCore_Tests {
         SMSCore.sendMessage(messageTest, null, null);
         fail();
     }
-
 }
