+++ webdictionary/src/androidTest/java/com/example/webdictionary/NetworkConnection_Tests.java
@@ -16,7 +16,7 @@ import static org.junit.Assert.*;
  * @see <a href="http://d.android.com/tools/testing">Testing documentation</a>
  */
 @RunWith(AndroidJUnit4.class)
-public class ExampleInstrumentedTest {
+public class NetworkConnection_Tests {
     @Test
     public void useAppContext() {
         // Context of the app under test.
